requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
asyncio==3.4.3
attrs==24.2.0
Authlib==1.3.2
av==13.0.0
bcrypt==4.0.1
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
cryptography==43.0.1
cssselect==1.2.0
cssutils==2.11.1
cymem==2.0.8
daily-python==0.11.0
debugpy==1.8.6
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
email-validator==2.1.0.post1
emails==0.6
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.109.1
filelock==3.16.1
firecrawl==1.2.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.8.1
grpcio==1.66.1
grpcio-health-checking==1.66.1
grpcio-tools==1.66.1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langcodes==3.4.1
language_data==1.2.0
livekit==0.17.4
livekit-agents==0.10.0
livekit-api==0.7.0
livekit-plugins-anthropic==0.2.1
livekit-plugins-deepgram==0.6.7
livekit-plugins-elevenlabs==0.7.5
livekit-plugins-openai==0.10.2
livekit-plugins-silero==0.7.1
livekit-protocol==0.6.0
lxml==5.3.0
Mako==1.3.5
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.19.2
openai==1.50.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
postgrest==0.16.11
premailer==3.10.0
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==5.9.8
psycopg==3.1.13
psycopg-binary==3.1.13
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.2.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
realtime==1.0.6
regex==2024.9.11
requests==2.32.3
retell-sdk==3.18.0
rich==13.8.1
safetensors==0.4.5
sentry-sdk==1.40.6
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
sqlmodel==0.0.21
srsly==2.4.8
stack-data==0.6.3
starlette==0.35.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
sympy==1.13.3
tenacity==8.2.3
termcolor==2.4.0
thefuzz==0.22.1
thinc==8.2.5
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
twilio==9.2.3
typer==0.9.4
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.24.0.post1
uvloop==0.20.0
validators==0.33.0
vapi-python==0.1.9
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.3.4
weaviate-client==4.7.1
websockets==12.0
wrapt==1.16.0
yarl==1.13.1
